refactor(Testing): Move a few image tests from WebGL to Core
Moving more tests to Core so they can be used for both WebGL and
WebGPU
