+++ Sources/Rendering/Core/ImageMapper/test/testImage.js
@@ -4,15 +4,15 @@ import testUtils from 'vtk.js/Sources/Testing/testUtils';
 import vtkImageGridSource from 'vtk.js/Sources/Filters/Sources/ImageGridSource';
 import vtkImageMapper from 'vtk.js/Sources/Rendering/Core/ImageMapper';
 import vtkImageSlice from 'vtk.js/Sources/Rendering/Core/ImageSlice';
-import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
+import 'vtk.js/Sources/Rendering/Misc/RenderingAPIs';
 import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
 import vtkRenderWindow from 'vtk.js/Sources/Rendering/Core/RenderWindow';
 
 import baseline from './testImage.png';
 
-test.onlyIfWebGL('Test ImageMapper', (t) => {
+test('Test ImageMapper', (t) => {
   const gc = testUtils.createGarbageCollector(t);
-  t.ok('rendering', 'vtkOpenGLImageMapper testImage');
+  t.ok('rendering', 'vtkImageMapper testImage');
 
   // Create some control UI
   const container = document.querySelector('body');
@@ -54,8 +54,8 @@ test.onlyIfWebGL('Test ImageMapper', (t) => {
   // modify objects in your browser's developer console:
   // -----------------------------------------------------------
 
-  // create something to view it, in this case webgl
-  const glwindow = gc.registerResource(vtkOpenGLRenderWindow.newInstance());
+  // create something to view it
+  const glwindow = gc.registerResource(renderWindow.newAPISpecificView());
   glwindow.setContainer(renderWindowContainer);
   renderWindow.addView(glwindow);
   glwindow.setSize(400, 400);
@@ -64,7 +64,7 @@ test.onlyIfWebGL('Test ImageMapper', (t) => {
     testUtils.compareImages(
       image,
       [baseline],
-      'Rendering/OpenGL/ImageMapper',
+      'Rendering/Core/ImageMapper',
       t,
       1,
       gc.releaseResources
