+++ Sources/Rendering/Core/ImageMapper/test/testImageColorTransferFunction.png
