+++ Sources/Rendering/Core/ImageMapper/test/testImageNearestNeighbor.png
