+++ Sources/tests.js
@@ -56,6 +56,9 @@ import './Rendering/Core/Glyph3DMapper/test/testGlyph3DMapper';
 import './Rendering/Core/HardwareSelector/test/testHardwareSelector';
 import './Rendering/Core/HardwareSelector/test/testHardwareSelectorGlyph';
 import './Rendering/Core/HardwareSelector/test/testHardwareSelectorSpeed';
+import './Rendering/Core/ImageMapper/test/testImage';
+import './Rendering/Core/ImageMapper/test/testImageNearestNeighbor';
+import './Rendering/Core/ImageMapper/test/testImageColorTransferFunction';
 import './Rendering/Core/Mapper/test/testEdgeVisibility';
 import './Rendering/Core/Mapper/test/testVectorComponent';
 import './Rendering/Core/PointPicker/test/testPointPicker';
@@ -65,10 +68,7 @@ import './Rendering/Core/SphereMapper/test/testDisableScalarColoring';
 import './Rendering/Core/SphereMapper/test/testSphere';
 import './Rendering/Core/StickMapper/test/testStick';
 import './Rendering/Misc/GenericRenderWindow/test/testGenericRenderWindowCreateDelete';
-import './Rendering/OpenGL/ImageMapper/test/testImage';
 import './Rendering/OpenGL/ImageMapper/test/testImageCroppingPlanes';
-import './Rendering/OpenGL/ImageMapper/test/testImageNearestNeighbor';
-import './Rendering/OpenGL/ImageMapper/test/testImageColorTransferFunction';
 import './Rendering/OpenGL/PolyDataMapper/test/testAddShaderReplacements';
 import './Rendering/OpenGL/PolyDataMapper/test/testClearShaderReplacements';
 import './Rendering/OpenGL/PolyDataMapper/test/testClippingPlanes';
